refactor: edit Routing and Redirection
