+++ src/components/layout/Layout.tsx
@@ -12,14 +12,16 @@ export default function Layout(){
 };
 
 const Home = () => {
-    let auth = useAuth();
+    let { authInfo, signOut } = useAuth();
     let navigate = useNavigate();
+
+    console.log(authInfo);
     
     return (
         <>
             <div className="flex flex-col justify-center items-center place-items-center h-screen">
                 <h1 className="font-semibold mb-5 text-xl">To Do</h1>
-                    {!auth.authInfo ? (
+                    {!authInfo ? (
                         <>
                             <button className="my-0.5 w-60 text-base py-2 bg-green-1 text-white rounded-2xl">
                                 <Link to="/signin">로그인</Link>
@@ -30,16 +32,14 @@ const Home = () => {
                         </>
                     ) : (
                         <>
-                        <button className="bg-green-1 text-white my-0.5 w-60 text-base py-2 rounded-2xl hover:bg-green-2"
-                            onClick={() => {
-                                navigate("/todo");
-                            }}
-                            >
+                        <Link to="/todo">
+                        <button className="bg-green-1 text-white my-0.5 w-60 text-base py-2 rounded-2xl hover:bg-green-2">
                             My To Do
                         </button>
+                        </Link>
                         <button className="bg-green-1 text-white my-0.5 w-60 text-base py-2 rounded-2xl hover:bg-green-2"
                             onClick={() => {
-                            auth.signOut(() => navigate("/signin"));
+                                signOut(() => navigate("/signin"));
                             }}
                             >
                             로그아웃
